+++ packages/plugin-starknet/src/index.ts
@@ -1,6 +1,23 @@
 import { Plugin } from "@ai16z/eliza";
 import { executeSwap } from "./actions/swap";
 
+export const PROVIDER_CONFIG = {
+    BIRDEYE_API: "https://public-api.birdeye.so",
+    MAX_RETRIES: 3,
+    RETRY_DELAY: 2000,
+    DEFAULT_RPC: "https://api.mainnet-beta.solana.com",
+    TOKEN_ADDRESSES: {
+        SOL: "So11111111111111111111111111111111111111112",
+        BTC: "qfnqNqs3nCAHjnyCgLRDbBtq4p2MtHZxw8YjSyYhPoL",
+        ETH: "7vfCXTUXx5WJV5JADk17DUJ4ksgau7utNKj4b963voxs",
+        Example: "2weMjPLLybRMMva1fM3U31goWWrCpF59CHWNhnCJ9Vyh",
+    },
+    TOKEN_SECURITY_ENDPOINT: "/defi/token_security?address=",
+    TOKEN_TRADE_DATA_ENDPOINT: "/defi/v3/token/trade-data/single?address=",
+    DEX_SCREENER_API: "https://api.dexscreener.com/latest/dex/tokens/",
+    MAIN_WALLET: "",
+};
+
 export const starknetPlugin: Plugin = {
     name: "starknet",
     description: "Starknet Plugin for Eliza",
